networking refactor: more robustness for indirect nodes (#106)
This PR refactors networking to be faster and less buggy, also, adds keepalive between indirect nodes and their routers.

This is a breaking change to "networking protocol" so I will update the rolr's to match. Therefore when this is merged, all running nodes need to start using this version of networking code to get routed.